% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/habitat_suitability.R
\name{burp_func}
\alias{burp_func}
\title{Webscrape BURP}
\usage{
burp_func(id)
}
\arguments{
\item{id}{A character of a BURP id.}
}
\value{
A tibble with water quality information.
}
\description{
Webscrape BURP
}
