% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/habitat_suitability.R
\name{get_burp}
\alias{get_burp}
\title{Title}
\usage{
get_burp(years = c(1994:2008, 2010:2023), huc12)
}
\arguments{
\item{years}{A vector of years to query BURP with.}

\item{huc12}{A sf object with USGS HUC 12s.}
}
\value{
A sf object.
}
\description{
This function webscrapes https://www2.deq.idaho.gov/water/BurpViewer/ website to get relevant water quality data (temperature, ph, do, etc). It is
recommended to do this in parallel because it will take a really long time to get all of the data!
}
