% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/retrieval.R
\name{get_data}
\alias{get_data}
\title{Get Data}
\usage{
get_data(aoi)
}
\arguments{
\item{aoi}{A sf object.}
}
\value{
A list of data used in the model; NHPLus Flowlines and Waterbodies, Water Quality Portal, Norwest data
}
\description{
Get Data
}
