% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/final_scores.R
\name{get_final_scores}
\alias{get_final_scores}
\title{Get Final AIS Risk Scores}
\usage{
get_final_scores(link_wqs_with_huc12)
}
\arguments{
\item{link_wqs_with_huc12}{A previously created \code{link_wqs_with_huc12()} object.}
}
\description{
Get Final AIS Risk Scores
}
