% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/social_suitability.R
\name{get_mussel_proximity}
\alias{get_mussel_proximity}
\title{Get Mussel Proximity}
\usage{
get_mussel_proximity(
  boundary,
  non_usfs_roads,
  usfs_roads,
  mussel_locations,
  huc12,
  conda_path =
    "C:\\\\Program Files\\\\ArcGIS\\\\Pro\\\\bin\\\\Python\\\\envs\\\\arcgispro-py3",
  workspace = getwd()
)
}
\arguments{
\item{boundary}{A sf object for clipping the area of interest.}

\item{non_usfs_roads}{A sf object with non USFS roads.}

\item{usfs_roads}{A sf object with USFS roads.}

\item{mussel_locations}{A sf object with Mussel Locations.}

\item{huc12}{A sf object with USGS HUC 12s.}

\item{conda_path}{A file path to conda env, e.g. \code{r'{C:\\Program Files\\ArcGIS\\Pro\\bin\\Python\\envs\\arcgispro-py3}'}.}

\item{workspace}{A local file path to store file with \code{arcpy}.}
}
\value{
A tibble with mussel proximity by meters \code{mean_dist} and HUC 12.
}
\description{
Get Mussel Proximity
}
