% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/habitat_suitability.R
\name{get_streamtemp}
\alias{get_streamtemp}
\title{Get Stream Temperature}
\usage{
get_streamtemp(boundary, processing_units)
}
\arguments{
\item{boundary}{A sf object for clipping the area of interest.}

\item{processing_units}{A character of the Processing Units, e.g. 'SpoKoot'.}
}
\description{
Get Stream Temperature
}
\note{
This is only for flowlines and not waterbodies.
}
