% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/social_suitability.R
\name{get_type_size_position}
\alias{get_type_size_position}
\title{Get Type, Size, Position}
\usage{
get_type_size_position(waterbodies, flowline, elevation = 1500)
}
\arguments{
\item{elevation}{A numeric. Used for waterbody upland/lowland filtering, in meters.}

\item{wb_layer}{A sf object of NHDPlus V2 waterbodies.}

\item{flowline_layer}{A sf object of NHDPLus V2 flowlines.}
}
\description{
Get Type, Size, Position
}
