% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/social_suitability.R
\name{link_mp_with_huc12}
\alias{link_mp_with_huc12}
\title{Connect Lakes with Outlet COMID}
\usage{
link_mp_with_huc12(mussel_proximity_df, type_size_position, huc12)
}
\arguments{
\item{mussel_proximity_df}{A previously created \code{get_mussel_proximity()} object.}

\item{type_size_position}{A previously created \code{get_type_size_position()} object.}

\item{huc12}{A sf object with USGS HUC 12s.}
}
\description{
Connect Lakes with Outlet COMID
}
