% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/habitat_suitability.R
\name{link_wqs_with_huc12}
\alias{link_wqs_with_huc12}
\title{Link Water Quality Attributes}
\usage{
link_wqs_with_huc12(wqp, burp, streamtemp, waterbodytemp, link_mp_with_huc12)
}
\arguments{
\item{wqp}{A previously created \code{get_wqp()} object.}

\item{burp}{A previously created \code{get_burp()} object.}

\item{streamtemp}{A previously created \code{get_streamtemp()} object.}

\item{waterbodytemp}{A previously created \code{get_waterbodytemp()} object.}

\item{link_mp_with_huc12}{A previously created \code{link_mp_with_huc12()} object.}
}
\description{
This functions puts the pieces together for different water quality attributes. It uses a lot of joins and selecting so
syntax is very important.
}
